refactor: types and misc
